+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Added `--clean` option to barrel script
+- Added `useOverwritableState` to wrap `useState` and `useEffect` in uncontrolled components
 
 ### Changed
 - Changed `barrel` script location to a dedicated [scripts folder](scripts)